feat: add create and edit pages

+++ frontend/src/app/home/home.ts
@@ -1,6 +1,7 @@
 import { CommonModule } from '@angular/common';
 import { Component } from '@angular/core';
 import { FormsModule } from '@angular/forms';
+import { Router, RouterLink } from '@angular/router';
 import { BookService } from './api/book/service/book.service';
 import { CustomerService } from './api/customer/service/customer.service';
 import { PublisherService } from './api/publisher/service/publisher.service';
@@ -14,7 +15,7 @@ type CustomerPublisherFilter = 'ALL' | 'WITHOUT_BORROWED_BOOKS' | 'WITH_BORROWED
 
 @Component({
   selector: 'app-home',
-  imports: [ FormsModule, CommonModule],
+  imports: [ FormsModule, CommonModule, RouterLink],
   templateUrl: './home.html',
   styleUrl: './home.scss',
 })
@@ -35,13 +36,15 @@ export class Home {
   constructor(
     private bookService: BookService,
     private customerService: CustomerService,
-    private publisherService: PublisherService
+    private publisherService: PublisherService,
+    private router: Router
   ){}
 
   ngOnInit(){
     this.findAllBooksByName();
   }
 
+  //menu
   selectMenu(menu: Menu){
     this.selectedMenu = menu;
     this.searchSelectedMenu();
@@ -59,6 +62,34 @@ export class Home {
     }
 
     this.findAllPublishersByName();
+  }
+
+  get createLabel(): string {
+    if(this.selectedMenu === 'BOOKS'){
+      return 'New Book';
+    }
+
+    if(this.selectedMenu === 'CUSTOMER'){
+      return 'New Customer';
+    }
+
+    return 'New Publisher';
+  }
+
+  createSelectedMenu(){
+    this.router.navigate([this.createRoute()]);
+  }
+
+  createRoute(): string {
+    if(this.selectedMenu === 'BOOKS'){
+      return '/books/create';
+    }
+
+    if(this.selectedMenu === 'CUSTOMER'){
+      return '/customers/create';
+    }
+
+    return '/publishers/create';
   }
 
   selectBookFilter(filter: BookFilter){
@@ -92,6 +123,9 @@ export class Home {
   }
 
   deleteBook(id: number){
+
+    if(!window.confirm("Delete this Book?")) return;
+
     this.bookService.delete(id).subscribe({
       next: () => {
         this.books = this.books.filter((book) => 
@@ -124,6 +158,9 @@ export class Home {
   }
 
   deleteCustomer(id: number){
+
+    if(!window.confirm("Delete this Customer?")) return;
+
     this.customerService.delete(id).subscribe({
       next: () => {
         this.customers = this.customers.filter((customer) => 
@@ -151,6 +188,9 @@ export class Home {
   }
 
   deletePublisher(id: number){
+
+    if(!window.confirm("Delete this Publisher?")) return;
+
     this.publisherService.delete(id).subscribe({
       next: () => {
         this.publishers = this.publishers.filter((publisher) => 
@@ -162,5 +202,5 @@ export class Home {
       }
     });
   }
-  
+
 }
